requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-filter==24.3
django-rest-passwordreset==1.4.0
drf-yasg==1.21.8
PyYAML==6.0.2
requests==2.32.3
Pillow==11.0.0
python-dotenv==1.0.1